Fixed|libdeng1|Reader: Reading 16/32-bit values
An incorrect return type for Reader_8() caused the higher bytes of
16/32-bit values to be truncated to zero.

+++ doomsday/libdeng1/src/reader.c
@@ -31,9 +31,9 @@
 #  define Reader_TypeCheck(r, code)
 #endif
 
-#define Reader_16(reader)        (Reader_8(reader,  0) | Reader_8(reader,  8))
-#define Reader_32(reader)        (Reader_8(reader,  0) | Reader_8(reader,  8) | \
-                                  Reader_8(reader, 16) | Reader_8(reader, 24))
+#define Reader_16(reader)        ( (uint16_t) (Reader_8(reader,  0) | Reader_8(reader,  8)) )
+#define Reader_32(reader)        ( Reader_8(reader,  0) | Reader_8(reader,  8) | \
+                                   Reader_8(reader, 16) | Reader_8(reader, 24) )
 
 typedef struct readerfuncs_s {
     Reader_Callback_ReadInt8  readInt8;
@@ -53,7 +53,7 @@ struct reader_s
     readerfuncs_t func;
 };
 
-static uint8_t Reader_8(Reader *reader, int shift)
+static uint32_t Reader_8(Reader *reader, int shift)
 {
     return reader->data[reader->pos++] << shift;
 }
